+++ crates/ruma-client-api/src/appservice.rs
@@ -1,3 +1,4 @@
 //! Endpoints part of the application service extension of the client-server API
 
+pub mod request_ping;
 pub mod set_room_visibility;