+++ crates/ruma-client-api/CHANGELOG.md
@@ -12,6 +12,7 @@ Improvements:
 - Add support for using an existing session to log in another (MSC3882 / Matrix 1.7)
 - Add support for media download redirects (MSC3860 / Matrix 1.7)
 - Stabilize support for asynchronous media uploads (MSC2246 / Matrix 1.7)
+- Add support for the appservice ping mechanism (MSC 2659 / Matrix 1.7)
 
 # 0.16.2